requirements.txt
RPi.GPIO
Adafruit_DHT
gspread
oauth2client
Flask
smtplib
